multi_db/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.1.13
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.5.0
mongoengine==0.27.0
psycopg2-binary==2.9.9
pymongo==3.12.3
pytz==2024.1
sqlparse==0.2.4
typing_extensions==4.9.0
